feat(builder): merge lazy tests into BuilderTest
