--- tests/resources/views/pages/builder-lazy-browser-test.blade.php
@@ -1,7 +1,0 @@
-<x-filament-panels::page>
-    <form wire:submit="save">
-        {{ $this->form }}
-
-        <x-filament::button type="submit">Save</x-filament::button>
-    </form>
-</x-filament-panels::page>
